+++ src/pages/User.tsx
@@ -65,8 +65,6 @@ const User = () => {
   const [auth, setAuth] = useState<Payload>(initPayload);
   const [state, setState] = useState(userState.user);
   const [group, setGroup] = useState<Array<groupid>>(initGroups);
-  const [following, setFollowing] = useState();
-  const [follower, setFollower] = useState();
 
   useEffect(() => {
     //get user auth
@@ -109,16 +107,6 @@ const User = () => {
     partyAxios.getJoinedParty(param.userId).then((res) => {
       setGroup(res.data.groups);
     });
-
-    //set FollowList
-    userAxios.getFollowUser(param.userId, "follower").then((res) => {
-      console.log("팔로우", res.data);
-      setFollower(res.data);
-    });
-    userAxios.getFollowUser(param.userId, "following").then((res) => {
-      console.log("팔로잉", res.data);
-      setFollowing(res.data);
-    });
   }, []);
 
   // paint user proflie or otherUser Profile
@@ -136,8 +124,6 @@ const User = () => {
           auth={auth}
           user={state}
           uid={param.userId}
-          follower={follower}
-          following={following}
         ></MyInfo>
       </React.Fragment>
     );
